[Feat] AttendeeRule protocol 추가, Attendee 구조체가 AttendeeRule 프로토콜 채택
