+++ LuckyCardGame/LuckyCardGame.xcodeproj/xcuserdata/eunno.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -31,8 +31,8 @@
             filePath = "LuckyCardGame/Model/LuckyCardGame.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "32"
-            endingLineNumber = "32"
+            startingLineNumber = "34"
+            endingLineNumber = "34"
             landmarkName = "gameStart(attendeeNum:)"
             landmarkType = "7">
             <Actions>
@@ -358,22 +358,6 @@
             startingLineNumber = "33"
             endingLineNumber = "33"
             landmarkName = "init(frame:cardAnimal:cardNumber:)"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "1B431DB1-6139-40BB-A4FB-A903F2014AF4"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "LuckyCardGame/Model/Attendee.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "93"
-            endingLineNumber = "93"
-            landmarkName = "init(deck:)"
             landmarkType = "7">
          </BreakpointContent>
       </BreakpointProxy>
