+++ LuckyCardGame/LuckyCardGame/Model/AboutCard/LuckyCard.swift
@@ -16,7 +16,7 @@ protocol Card{
 /*
  MARK: LuckyCard하나의 정보를 표현한 struct입니다. set이 없기 때문에 굳이 class로 만들 이유가 없다고 생각해서 struct를 사용했습니다.
  */
-struct LuckyCard: Card, Hashable, Equatable{
+class LuckyCard: Card, Equatable{
     
     static func == (lhs: LuckyCard, rhs: LuckyCard) -> Bool {
         return lhs.number == rhs.number && lhs.animal == rhs.animal
@@ -25,6 +25,7 @@ struct LuckyCard: Card, Hashable, Equatable{
     private(set) var number: CardNumber
     private(set) var animal: Animal
     private(set) var isFlipped: Bool
+    
     var describe: String{
         return "\(animal.rawValue)" + String(format:"%02d", number.rawValue)
     }
@@ -35,15 +36,15 @@ struct LuckyCard: Card, Hashable, Equatable{
         self.isFlipped = isFlipped
     }
     
-    public mutating func setNumber(number: CardNumber){
+    public func setNumber(number: CardNumber){
         self.number = number
     }
     
-    public mutating func setAnimal(animal: Animal){
+    public func setAnimal(animal: Animal){
         self.animal = animal
     }
     
-    public mutating func flipCard(){
+    public func flipCard(){
         self.isFlipped = !self.isFlipped
     }
 }
